Current iteration of the solver.
There are still a number of logical rules I can implement with the hope of reducing the reliance on backtracking. Right now, it handles 40 by 30 boards fairly well, but it appears to struggle with 50 by 40 boards. I suspect that there may be a perverse scaling effect going on because of the way that I search for the next edge to trying filling in. It's equally plausible that the larger boards are genuinely becoming intractable, because this problem is known to be NP-complete.